Rename ProgramError2 files to ProgramError
