+++ packages/voictents-and-estinants-engine/src/custom/programmable-units/engine-program/getEngineProgramLocator.ts
@@ -61,7 +61,7 @@ import {
   GenericProgramError2Voque,
   ReportedProgramError2,
   ReportingEstinantLocator,
-} from '../error/programError2';
+} from '../error/programError';
 
 const ESTINANT_NAME = 'getEngineProgramLocator' as const;
 type EstinantName = typeof ESTINANT_NAME;
