+++ packages/voictents-and-estinants-engine/src/custom/programmable-units/error/signalError.ts
@@ -2,7 +2,7 @@ import { buildEstinant } from '../../adapter/estinant-builder/estinantBuilder';
 import {
   GenericProgramError2Voque,
   PROGRAM_ERROR_2_GEPP,
-} from './programError2';
+} from './programError';
 
 /**
  * Sets a non-zero exit code if there is a program error.
