+++ packages/voictents-and-estinants-engine/src/custom/programmable-units/type-script-file-relationships/graph-element/getFileNodeMetadata.ts
@@ -5,7 +5,7 @@ import {
   GenericProgramError2Voque,
   ReportedProgramError2,
   ReportingEstinantLocator,
-} from '../../error/programError2';
+} from '../../error/programError';
 import { Shape } from '../../graph-visualization/directed-graph/attribute';
 import {
   TYPE_SCRIPT_FILE_GEPP,
