+++ packages/voictents-and-estinants-engine/src/custom/programmable-units/type-script-file/getTypeScriptFileImportList.ts
@@ -18,7 +18,7 @@ import {
   GenericProgramError2Voque,
   ReportedProgramError2,
   ReportingEstinantLocator,
-} from '../error/programError2';
+} from '../error/programError';
 
 const ESTINANT_NAME = 'getTypeScriptFileImportList' as const;
 type EstinantName = typeof ESTINANT_NAME;
