+++ packages/voictents-and-estinants-engine/src/custom/programs/categorize-files/categorizeFiles.ts
@@ -9,7 +9,7 @@ import { enumerateFileSystemObjects } from '../../programmable-units/file/enumer
 import { InMemoryVoictent } from '../../../core/engine/inMemoryVoictent';
 import { assertFileExtensionIsKnown } from './assertFileExtensionIsKnown';
 import { ProgramFileCache } from '../../../utilities/programFileCache';
-import { ProgramError2Voictent } from '../../programmable-units/error/programError2Voictent';
+import { ProgramError2Voictent } from '../../programmable-units/error/programErrorVoictent';
 import { reportErrors } from '../../programmable-units/error/reportErrors';
 import { signalError } from '../../programmable-units/error/signalError';
