+++ packages/voictents-and-estinants-engine/src/custom/programs/model-ci/assertCiModelHasAllPrograms.ts
@@ -11,7 +11,7 @@ import {
   GenericProgramError2Voque,
   ReportedProgramError2,
   ReportingEstinantLocator,
-} from '../../programmable-units/error/programError2';
+} from '../../programmable-units/error/programError';
 
 const ESTINANT_NAME = 'assertCiModelHasAllPrograms' as const;
 type EstinantName = typeof ESTINANT_NAME;
